GIX-1363: Register Vote through GovernanceApiService (#2053)
# Motivation

The fix of the voting button used a test feature.

This is the fix we intended. Calls to governance canister should be done
through the governance api service and this one decides when to reset or
not the cache.

# Changes

* Add method "registerVote" in the GovernanceApiService. It clears the
cache after it.
* Move "registerVote" api function to "governace.api". Slowly, start
implementing the agreed pattern that api should be grouped by canister.
* Voting service uses governance api service instead of the api function
directly. That means we can remove the call to `resetNeuronsApiService`.

# Tests

* Move test for the api function "registerVote" to the
"governace.api.spec".
* Add test for registerVote in the governance api service.
* Change mocks due to moving api function to another file.