Fix: SNS Wallet page infinite loading (#2056)
# Motivation

Deep link to an SNS Wallet wasn't working.

The problem was that the `load` function that sets the context store was
not called after the accounts were loaded in the store.

It was called only when the identifier was there.

It worked with normal navigation because the store was already filled
with the accounts. But if the user arrives with the deep link, the
accounts are loaded. Therefore, the `load` function was called before
they were loaded, but not after they were in the store.

# Changes

* Call `load` function also when sns accounts store changes in
SnsWallet.svelte page.

# Tests

* Upcoming in another PR along with other test for deep links.