Page object cleanups based on feedback from PR 2062 (#2068)
# Motivation

The getters on page objects that return page objects can be a bit
confusing, especially then the component name is in plural. For example
getNnsNeurons does not return multiple neurons but the page object for
the component NnsNeurons.

We also had to add the same css to a number of different components just
to be able to wrap it in an element with a test ID.

# Changes

1. For getters that return a page object, make the name end with "Po".
2. For getters that return a list of page objects, make the name end
with "Pos".
3. For getters that return whether a page object is non-nullish, make
the name end with "Po".
4. Add TestIdWrapper to wrap multiple element with a data-tid and apply
display:contents.

# Tests

1. `npm run test`
2. Manual testing.