GIX-1394: Invalid address depends on universe (#2060)
# Motivation

Address formats depends on universe. Currently, we were accepting a
valid ICRC in the NNS universe, and viceversa, a valid NNS ICP address
in SNS universe.

This lead to a bad UX when the user got an error message when trying to
send the tokens.

Solution: Use different validations depending on the universe.

# Changes

Major changes:
* Add "rootCanisterId" parameter to accounts util `invalidAddress` and
use it for the validation.
* Split `invalidICPOrIcrcAddress` into two `invalidIcpAddress` and
`invalidIcrcAddress`.
* Add prop "rootCanisterId" in component AddressInput.svelte and use it
for the validation.
* Pass prop "rootCanisterId" in TransactionForm until it reaches the
AddressInput.svelte.

Minor changes:
* Rename `Address.svelte` to `NnsAddress.svelte` to make clear where it
should be used. (This component might be removed when we refactor the
modals).
* Change usage of `invalidICPOrIcrcAddress` with the new helpers.

# Tests

* Test new functionality in AddressInput.svelte.
* Test `invalidAddress` new parameter.
* Test new accounts utils `invalidIcpAddress` and `invalidIcrcAddress`.