Downgrade upgrade test (#2067)
# Motivation
We are making more changes to the nns-dapp backend and in particular to
the stable memory. It would be good to have at least a very basic
downgrade-upgrade test in CI.

Requirements:
- Fast; should not slow down CI significantly.
- Verify that there are no panics in the pre and post-upgrade hooks.

Not covered:
- Detailed checks that might be applicable to a specific upgrade, e.g.
making sure that stable data is preserved or manipulated in a certain
way.

# Changes
- Add a script that performs a downgrade and then an upgrade. Given that
the current code is usually deployed when testing, a downgrade-upgrade
test is quicker to implement than an upgrade-downgrade test, which would
really be a downgrade-upgrade-downgrade-upgrade test.
- Perform a health check by calling the stats endpoint. This is much
faster than an e2e test.

# Tests
- No change to rust code: See how these tests perform in CI.
- Change to Rust code: I have run the script in this PR on code that
breaks that stable memory and indeed, the script fails.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>