clippy (aggregator): Prohibit panicking except where explicitly allowed (#2072)
# Motivation
Panicking can be anywhere between bad and disasterous if it happens in
the wrong place. Unfortunately, prohibiting panic on a function does not
prohibit panics in dependencies of that function. So, instead, we
prohibit panicking everywhere and whitelist a few cases where it is safe
and correct.

# Changes
In the aggregator:
- Move deny panic to the lib.rs module so that it covers the whole
codebase.
- Remove panics where possible, whitelist other cases.

# Tests
This is a static compile time check.  If it compiles, it works.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>