feat: read destination address with an integrated QR code reader (#2092)
# Motivation

Scan destination address with an integrated QR code reader.

# PRs

- [x] https://github.com/dfinity/gix-components/pull/171

# Notes

Because the web worker is bundled within the gix-cmp, both `npm run dev
/ test` are not able to loads it. This PR amend this limitation given
the fact that most important test on the QR code reader are performed
with e2e test in gix-component. We might try to find a solution in
further PRs, did not find despite having tried quite a lot...

# Changes

- integrate new `<QRReaderModal />` component
- add a button to open QR code in `AddressInput`
- implement a new step within `TransactionModal`