Add performance counters to state (#2115)
# Motivation
We wish to add performance counters to the state. This introduces a risk
that the post-upgrade hook will fail to recreate the state. We therefore
add this in an isolated PR with corresponding tests.

Note: The performance counters are NOT saved to stable memory as they
are reset on upgrade, so there is no point in saving them. The
information on this is a bit inconsistent but the spec matches the
implementation in this regard (but not in all regards):
https://internetcomputer.org/docs/current/references/ic-interface-spec#system-api-performance-counter
As such this PR _used_ to add a third candid field with performance data
but no more.

# Changes
- Add perf data to the state

# Tests
- Add a test that serialization and deserialisation of state works for
the relevant data.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>