Cache only docker main (#2124)
# Motivation
Improve docker caching

- GitHub has limited storage for caches. If there are too many branches
pushing changes to the cache, we get cache evictions and things run
slowly. Let's rather cache main only, and branches can copy from main
but not write docker caches.
- On the other hand, we can cache more steps, so a given PR rebuilds
only the parts of the codebase that have diverged from main.

# Changes
- Add a github action that saves a full docker build every time a commit
is merged into main.
- Make other docker builds read-only

# Future
- Make more extensive use of the docker builds in CI.

# Tests
- Timing
  - Verified that bumping this PR yields:
- Cached local docker build completes in about 30 seconds, most of which
is spent downloading the cache.
- Update: After several runs, it frequently takes around 30s but
sometimes it takes 1m, doing nothing but downloading the docker layers.
- Full mainnet docker build completes in just under 5 minutes; the
builder stage hits the cache, frontend and nns-dapp must be rebuilt.
- For a native build, installing the tooling, without building, takes
just under 5 minutes. Build is extra but hard to measure as it is not a
separate step.
- Note: I have seen e2e tests flake sometimes. Should scrutinize them
closely and fix.

# Future
- Use docker builds for the aggregator and e2e test workflows.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>