Chore: Fix flaky test (#2148)
# Motivation

There was a flaky test that would fail unexpectedly.

# Changes

* Fix the test in sns sale api spec. The expectation compares to what
the mocked method returns.

The reason it was flaky was because it was comparing to another ticket
created at the beginning of the test. Creating this test tickets uses
Date.now(), which meant that it was comparing two different tickets
created with the same parameters, but because of Date.now, they could be
different tickets.

# Tests

Only test changes.