Upgrade instruction counter (#2064)
# Motivation
If the post-upgrade hook consumes too many cycles, it may fail. Such
failures are irrecoverable except by rolling back the subnet. This is
bad. We would like to have more visibility into cycle consumption.

Cycle consumption is instruction count times a factor. The factor
depends on such things as subnet size. As such,
cycle count is actually not useful to test in CI but it can be converted
into an instruction count limit that _can_ be tested in CI.

A good first step is to capture instruction counts and make them
available for examination.

# Changes
- Extend the existing `get_stats()` API method to include instruction
counters in the response.
- Define an instruction counter struct including the count, time, and a
text label.
  - Add a circular buffer with the instruction counts to stable memory.
- Add a script to get the last upgrade instruction count.

# Future
- Determine an instruction count threshold for the post upgrade hook
- Add the post-upgrade instruction count to the metrics.
- Add methods to get the cycle count.
- Take measures to  reduce the post-upgrade cycle count.
Note: Measurements show that the overwhelming majority of cycles are
consumed when unpacking and re-certifying assets. Adding the assets
could be done in a separate function call after the post-upgrade hook,
but triggered by the post-upgrade.

# Tests
- In CI, upgrade and get the instructions consumed in the last upgrade.
- Note: This test does NOT currently verify that the instruction count
is below some threshold. Once such a threshold is established, this
would be the right place to enforce it.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>