Make PageObjectElement.getText() async (#2154)
# Motivation

We want to share page objects between unit tests and e2e tests.
In e2e tests some things have to be async because they involve making a
request to the browser driver.
So for compatibility they have to be async in the platform agnostic
interface.
One such things is getting the text content of an element.

# Changes

1. Make PageObjectElement.getText() async.
2. Make everything derived from it async as well.

# Tests

npm run test
npm run check