Allow SNS governance API fake to be paused (#2168)
# Motivation

Sometimes we want to test a component in a loading state.
For the component to be in this state, often an API call needs to not be
resolved yet.
Tests might rely on promises not resolving instantly but this is
brittle.
Better is to explicitly only resolve the promise after the expectations
are made.
A nice way to do this is to put a fake API in a "paused" state until the
expectation have been made.

# Changes

1. Allow SNS governance API fake to be paused.
2. Pause the SNS governance API in
`frontend/src/tests/lib/pages/SnsProposalDetail.spec.ts` to make the
test less brittle. Otherwise it would fail with PR #2159.

# Future

Other fakes can be done in separate PRs.

# Tests

I tested the code in PR #2159 and tested that it makes the test
consistent even with many extra `await null`s.