Docker build e2e (#2178)
# Motivation
Faster CI. We already have builds cached in docker. Reusing the cache
means that we don't need to rebuild unless we need to. Furthermore, even
on cache miss, a docker build takes about 4 minutes which is about as
long as it takes just to install the native build tools without even
starting to build.

# Changes
- docker-build the nns-dapp
- Don't build locally
- Remove unneeded tool installation and an unused env var.

# Tests
See CI

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>