Script to run snsdemo state for local development (#2222)
# Motivation

For local development it's useful to quickly run a local replica with a
certain state.
The snsdemo repo has a script to install SNS projects and a script to
save the replica state.
We use this state to run e2e tests and we can use the same state for
local development.

# Changes

Add `scripts/dev-local-state.sh` which:
1. Takes a zip file as prepared by snsdemo/bin/dfx-state-save
2. Extracts it
3. Moves existing state to a backup location
4. Moves the extracted state to the locations where it will be used by
`dfx start`
5. Runs `dfx start`
6. Runs `npm run dev`

When the dev server exists:
1. Stops the local replica
2. Restores the backed up state

# Tests

1. Manually created state with snsdemo/bin/dfx-state-save after running
the demo.
2. Used the script to run the state.

Sadly, using the state produced on the GitHub runner does not work on my
M1 Mac, even when created on a Mac runner. It may or may not work on an
Intel Mac.